22-03-08 11:29  validator form update trick in progress

+++ src/Controller/FigureController.php
@@ -399,36 +399,26 @@ class FigureController extends AbstractController
             try{
         
                 $descriptionTrick = $formDescriptionTrick->getData();
-                // dump($descriptionTrick);
-                // exit;
+
                 $nameTrickField = $descriptionTrick->getName();
 
-                $nameExist = $figureRepository->findOneByName($nameTrickField);
- 
-                if (!$nameExist){
+                $descriptionfield = $descriptionTrick->getDescription();
+                $figureGroupSelect = $descriptionTrick->getFigureGroup();
+                $coverImageTrick = $descriptionTrick->getCoverImage();
 
-                    $descriptionfield = $descriptionTrick->getDescription();
-                    $figureGroupSelect = $descriptionTrick->getFigureGroup();
-                    $coverImageTrick = $descriptionTrick->getCoverImage();
+                $coverImageTrick == null ? 'defaultCoverImage' : $coverImageTrick;
 
-                    $coverImageTrick == null ? '' : $coverImageTrick;
-
-                    $nameTrickSluger = $slugger->slug($nameTrickField);
+                $nameTrickSluger = $slugger->slug($nameTrickField);
     
-            
-    
-                            $figure->setName($nameTrickField);
-                            $figure->setSlug($nameTrickSluger);
-                            $figure->setDescription($descriptionfield);
-                            $figure->setCoverImage($coverImageTrick);
-                            $figure->setFigureGroup($figureGroupSelect);
-                            $this->entityManager->persist($figure);
-                            $this->entityManager->flush();
-
-                } else {
-                    $this->messageError  = " Attention ce nom existe déjà ! Veuillez changer l'intitulé";
-                    return $this->render('core/figures/trickEdit.html.twig', ['figure' => $figure, 'comments' => $comments, 'arrayMedias' => $arrayMedias, 'formDescriptionTrick' => $formDescriptionTrick->createView(),  'messageError' => $this->messageError ,'error' => false ]);
-                } 
+                $figure->setName($nameTrickField);
+                $figure->setSlug($nameTrickSluger);
+                $figure->setDescription($descriptionfield);
+                $figure->setCoverImage($coverImageTrick);
+                $figure->setFigureGroup($figureGroupSelect);
+                // dump($figure);
+                // exit;
+                $this->entityManager->persist($figure);
+                $this->entityManager->flush();
 
             }catch(Exception $e){
         
@@ -545,11 +535,18 @@ class FigureController extends AbstractController
  
         }
 
-            //delete coverImage
-            $fileNameCoverImage = $currentTrick->getCoverImage();
+
+        //delete coverImage
+
+        $fileNameCoverImage = $currentTrick->getCoverImage();
+
+        if($fileNameCoverImage !== 'defaultCoverImage' ) {
+
             $pathCoverImage = $this->getParameter('images_directory');
             $filePath = $pathCoverImage.'\\'. $fileNameCoverImage ;
             unlink($filePath);
+        }
+
